+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Add `is:retiredperk` search that highlights items that have a perk which can no longer drop for that item.
 * You can now click a Loadout name in Organizer's Loadouts column to quickly bring up this loadout for editing.
 * When hovering over subclass Aspects in Loadouts and Loadout Optimizer, the tooltip will now show the number of Fragment slots granted.
 * You can now bring up the Armory page for a weapon directly from the search bar by typing a weapon name there and clicking the corresponding entry.